Move `InMemoryFeedStore` to production code
